+++ nutils/topology.py
@@ -199,7 +199,7 @@ class Topology(types.Singleton):
     with matrix.Numpy():
       retvals = self.integrate([function.Inflate(function.asarray(func)[_], dofmap=ielem[_], length=len(self), axis=0) for func in funcs], **kwargs)
     retvals = [retval.export('dense') if len(retval.shape) == 2 else retval for retval in retvals]
-    return [function.elemwise(self.transforms, retval) for retval in retvals] if asfunction \
+    return [function.Elemwise(retval, self.f_index, dtype=float) for retval in retvals] if asfunction \
       else retvals
 
   @util.single_or_multiple